RISCV: Added riscv32 support to userlib
This adds the necessary code to make syscalls work on riscv32 targets.
It does this by moving architecture-specific code into the arch module,
which will select the appropriate implementation at compile-time. It
also adds logging via semihosting. With this commit, it is possible
to write apps for riscv32 targets.

NOTE: This has only been tested on riscv32imc and riscv32imac.

This commit is primarily based off of work by Jonathan Perkin.

Co-authored-by: Duncan Laurie <duncan@rivosinc.com>
Co-authored-by: Jonathan Perkin <jonathan@perkin.org.uk>