RISCV: Add support for riscv32 to test suite
This allows the test suite to run on riscv32 targets. The changes here
were fairly minimal, and include adding support for riscv semihosting,
using pseudo atomics where necessary, and changing tests that can be
used on riscv and removing the ones that cannot.

This commit is based off of work done by Jonathan Perkin.

Co-authored-by: Jonathan Perkin <jonathan@perkin.org.uk>