Revert "kern: remove klog!"
This reverts commit db991c0ccb98d1359e130f5f3fd13cb336b056d2.

Bring back klog. This is being done primarily for rv64. So, klog is
kept as a do-nothing for rv32 and arm. The reason for not enabling
klog support for rv32 and arm is because upstream and rivos/main
branches both do not have klog support and enabling klog means
slightly larger kernel sizes which will cause all apps/tests for
rv32 and arm to be updated in this branch. If we decide to enable klog
in rivos/main, we can go ahead and make the change in this branch too.

For rv64, there is value in having a way to dump kernel logs
especially during bring up. We might also want to consider having some
kind of klog even for production.

This change is mostly a partial revert of the change that removed
klog. Follow up commit adds klog support for rv64.