minimum self-signed DICE (#698)
* Update lpc55_support for compatibility with latest zeroize crate.

* lib-dice: Add new library with types for DICE artifacts and certs.

The current cert implementation is limited to the creation of a self
signed cert for the DeviceId & the Alias cert. Cert creation is as
simple as possible with a template for each cert (DER) that is populated
with fixed length data like the platform serial number, certificate
serial number, public key & signature. Future work will address
programatic generation of these templates at build time to replace the
currently static ones.

* lpc55/chip.toml: Add memory regions for DICE artifact handoff.

The memory used in this commit is SRAM intended for use by USB. The
choice of using this memory region was mostly driven by convenience and
pivoting to a different range should be straight forward. The size of
this region was chosen based on the estimated size of the data passed.
The DeviceId and Alias cert are each ~600 bytes and the seed required to
reconstruct the Alias key is 32 bytes. All of this fits comfortably in
2k.

* stage0: Use the dice library to implement minimal self-signed credentials.

This commit collects the DICE CDI from the NXP ROM, creates an ed25519
key pair from it (effectively the DeviceId) and uses this key to create
a self-signed DeviceId cert. The hash of the hubris image that will be
launched by stage0 is then hashed and the output used to generate the
Alias credentials (used for attestation). Finally the keying material
used to create the Alias key, and the cert chain from the Alias back to
the DeviceId are written to memory for use in an attestation exchange by
a future hubris task.

If run on an lpc55 with DICE disabled (the default config) none of the
DICE artifacts will be created and hubris will be booted as usual.