Added MutablePerfectHashMap and new functions
`MutablePerfectHashMap` is identical to `PerfectHashMap` except for the
fact that it has a new method: `set()`. This will write to the value
associated with the given key if there is already a value, otherwise
it will return an error.

This also adds the `contains()` method to both `PerfectHashMap` and
`MutablePerfectHashMap`. This method will indicate if there exists a
value for the given key.