Add a RISC-V PLIC driver along with API.
Currently, external interrupts on rv32 targets are handled by the
kernel, and when a peripheral has an interrupt it's assumed that it's an
external one. This, however, makes it impossible for interrupt vectors
>= 16, which are reserved for platform use, to be used. Additionally,
there's no way to support platforms using another external interrupt
controller. By moving the PLIC to userspace, the kernel can properly
service interrupts to the other lines, and any interrupt controller can
be plugged in and used.