mgmt-gateway: update host flash (#762)
* mgmt-gateway: update host flash

This tracks the corresponding changes to MGS messaging merged in
https://github.com/oxidecomputer/omicron/pull/1684:

1. `UpdateStart` has been broken into `UpdatePrepare` and
   `UpdatePrepareStatus` (which MGS will continue to send periodically
   until we respond that preparation is done), allowing for updates that
   have a potentially-long running prep step (like updating host flash,
   which can take up to several minutes to erase!).
2. Update messages now include a stream-id that we use to correlate
   related messages; we reject update messages that don't match our
   current stream ID.
3. Add handling for the new `UpdateAbort` abort message to cancel an
   in-progress update.

By far the most complex bit of this is 1: I've moved setting the system
timer out of `mgs_gimlet` (which previously set it only in relation to
flushing serial console uart packets out to MGS) and into `main`: now
the MGS handler only returns the deadline it wants `main` to set. If
we're in the process of prepping for a host flash update (i.e., we need
to erase the host flash), we'll set our deadline to 1 tick from now.
When it fires, we'll erase 8 sectors (takes about 1 second, worst case),
then return and allow `main` to check for other work. This allows us to
continue to be responsive to incoming notifications, importantly network
requests (allowing us to respond to the `UpdatePrepareStatus` messages
in a timely way!).