Adds build utils for peripheral information
Adds two functions to build_util: `task_peripherals()` and
`task_peripherals_str()`. The former returns a BTreeMap of the
peripheral names and a struct with their base address and size. The latter
takes that information and generates constants for them of the name
`{peripheral}_BASE_ADDR` and `{peripheral}_SIZE`, returning it as an
`std::String`.