sys/kern/rv64: Store current task pointer in mscratch
This change drops the global variable CURRENT_TASK_PTR and instead
stores the current task pointer in mscratch register. This drops
the usage of the variable in different functions and hence makes it
easier to split the files into separate submodules without having to
explicitly share this variable. This is in preparation to refactor
rv64 support into separate submodules in the future.

TEST=Verified that tests still pass for rv64.

Signed-off-by: Furquan Shaikh <furquan@rivosinc.com>