Dice cert builder + artifacts for sp-measure & rng task (#781)
* lib-dice: Implement builder pattern for certs.

This moves all of "setters" from the Cert trait into a new CertBuilder
trait. Implementations of the builder trait are added for the two
certs that we create. Builders are created by their respective 'new'
functions that take all required fields as parameters. Certs are then
instantiated by calling the 'sign' method on the builder.

* lib-dice: Create and transmit artifacts for the sp-measure task.

The cert created for the sp-meausre task is itself a CA cert like the
DeviceId. We'll use it to create credentials for the SP using the FWID
calculated for the SP image.

* lib-dice: Create and transmit artifacts to seed the lpc55 rng task.

This seed is handed off to a hubris task using a new memory range from the
same SRAM as the other DICE artifacts. This range is smaller than the
others because we don't pass any certs to the Rng driver.

* lib-dice: Create and transmit artifacts for the trust quorum dhe creds.

This is another key passed to the root of trust for reporting along with
the Alias key. The intended use for this key is described in RFD 238.

* lib-dice: Use static assert to ensure handling of handoff data is safe.

Additionally this commit converts constants for start & end of memory
ranges to use Range<usize>. Safety comments have also been updated to
better align with guidance from upstream.