Add extratext regions to task peripheral list
Apps can specify a section, `extratext`, that provides some extra space
for them to read/write in. These regions are used by adding them to the
`uses` field, but they come from the app.toml rather than the chip.toml.
There is nothing in our tree that uses this, but there is in upstream,
and without this change `xtask` will panic as it doesn't see the name in
the list of peripherals from the `chip.toml`.