sys/kern: Move `switch_to` to task.rs
This change moves `switch_to` to task.rs so that it can be used by
more than just syscalls. Also, this function is more related to tasks
than syscalls. This change is being done to make `switch_to` as the
common entrypoint for any kernel path to set next user task that
should be run. This helps avoid duplication of some common work that
needs to be done irrespective of architecture.