sys/kern: Use `switch_to` in `start_first_task`
This change uses the `switch_to` function in the startup path as well
since we were doing the same thing there w.r.t. applying memory
protection and setting current task pointer. One difference in
behavior with this change would be that `apply_memory_protection` is
called just before jumping to the first task which should be okay.

TEST=Verified that rv32 tests still pass.