riscv-plic-server: Get TaskIds using build_util
In order for the Plic server to be able to notify tasks of incoming
messages, a map of task names to their IDs needed to be exported in an
environment variable that was deserialized in the `build.rs`. Upstream
has recently added a function to `build_util` that provides this
information instead, so the PLIC server should be using that instead.