build: Generate irq notification constants
This adds a per-task environment variable, `HUBRIS_TASK_IRQS`, that
contains the code to create constants for interrupts the task recieves
through the kernel. For example, if a task has the interrupt
`"plic.irq" = 0x1`, the environemnt variable will contain:
`const PLIC_IRQ_NOTIFICATION: u32 = 0x1;`.

This also adds a function to `build_util`, `task_irq_consts()`, that
reads this environment variable and returns its contents as a `String`.