Standardize on a safe wrapper around std::env::var (#840)
This switches all uses of `std::env::var` to `build_env::env_var` and similar,
to ensure that we're properly tracking changes to environment variables (when
needed).

In the future, any use of `std::env::var` in build scripts should be treated
with a great deal of suspicion!

fixup: Fix build errors in `build-util`