xtask: Unified ELF32/64 Creation
In this patch, I make liberal use of rust's macro system and the
paste crate to create macros to generate both ELF32 and ELF64 headers.

In either case, we create both sets of containers (two vectors,
shared ELF header structure, shared section header string table).
However, only one of the [32|64] vector sets is filled, according
to the type of ELF to be generated. This is controlled by inspecting
arch_constants to discover if the end elf target is elf64.