Add graph module to xtask to create task priority graphs (#879)
* Add --graph $OUTPUT option to xtask dist to create task priority graph

A graph of tasks is created. Each task is labeled with its name and
priority in a box. Task slot relationships are represented by arrows.
Priority inversions are noted with a red arrow. Normal dependencies are
green arrows.

Use:

	cargo xtask graph -o graph.dot $APP
	# There are several utilities to work with Graphviz's files.
	# On Ubuntu:
	sudo apt install graphviz xdg-utils eog
	dot -Tsvg graph.dot
	xdg-open graph.svg

* Force row rank by priority and make graph its own module.

Also added the path to the app.toml file as a title in the graph.