xtask: make allowed-features list easier to maintain
Rather than a wrapped string constant that doesn't allow any internal
whitespace, the features are now listed in arrays and assembled into a
string when needed. This (IMO) makes it way easier to scan and
add/remove features.

NOTE: This has been slightly modified from the original commit, added in
https://github.com/oxidecomputer/hubris/pull/906. We have some extra
nightly features that we use, so they have been made a part of this
change as well.