ci: Remove Nix Flake Checks
While a theoretically awesome tool, Nix has demonstrated a pair of
extremely deleterious shortcomings that appear to be not considered
important by the Nix Community.

Specifically:
- Nix relies on the manual entry of input hashes for Cargo
  dependencies. This is extraordinarily cumbersome, and causes
  build failures on any dependency change. There was a suggestion
  to get this information automatically, but this has been stalled
  since 2019 (over 3 years, at the time of this commit):
  https://github.com/NixOS/nixpkgs/issues/63653

- Nix appears to be unable to handle multiple versions of Rust
  packages gracefully. The result is unexpected and uncommunicative
  failures, such as a failure to generate a symlink. This class of
  failure was first reported in 2017 (over 5 years ago, at the time
  of this commit):
  https://github.com/NixOS/nixpkgs/issues/22177

As a result, I do not believe that Nix can offer us a reliable
platform for CI of Hubris at this time. As observed on #46,
attempting to appease Nix can result in significant crate changes
to facilitate single revision jumps - consuming days of effort to
validate a change that is essentially operational.