Bump toolchain: 2021-11-01 (#923)
* sidecar-front-io: make enum Copy

Bumping the toolchain causes transcievers.rs to stop compiling because
this enum is consumed on two code paths, but is not marked Copy.
Deriving Copy insulates us from this.

I also added Debug because, well, it's useful.

* task/hiffy: fix warning in 1.65

* Bump toolchain: 2021-11-01

* Remove unused parens

NOTE: This is a slightly modified version of the commit from
https://github.com/oxidecomputer/hubris/pull/923. Changes were made to
accommodate the additions to `xtask` needed to support rv32 builds.

Co-authored-by: Matt Keeter <matt@oxide.computer>